+++ src/main/kotlin/com/github/kangwooc/dbmlplugin/DBMLIcons.kt
@@ -1,0 +1,29 @@
+package com.github.kangwooc.dbmlplugin
+
+import com.intellij.icons.AllIcons
+
+object DBMLIcons {
+    @JvmField
+    val FILE = AllIcons.FileTypes.Any_type
+    
+    @JvmField
+    val TABLE = AllIcons.Nodes.DataTables
+    
+    @JvmField
+    val COLUMN = AllIcons.Nodes.Field
+    
+    @JvmField
+    val ENUM = AllIcons.Nodes.Enum
+    
+    @JvmField
+    val PROJECT = AllIcons.Nodes.Project
+    
+    @JvmField
+    val TABLE_GROUP = AllIcons.Nodes.Folder
+    
+    @JvmField
+    val REF = AllIcons.Nodes.Related
+    
+    @JvmField
+    val INDEX = AllIcons.Nodes.KeymapAnt
+}